ball-machine/bridge/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyserial>=3.5
